Use an observer instead of a promise for a single action's results
Using an observer instead of a promise for a single
action's result allows a lot of code passing promises
through to be eliminated. It also simplified the
interaction for an end user of the API, as retrying a
failed request repeatedly no longer requires a potentially
endless chain of promises.