Minor API change, fixes to Actions, add method, and tests
Action result 'errors' really represent data conflicts
only, as other errors will still be handled through the
error callback of the action. The relevant flag has been
renamed to better reflect that.

Action observers should only call `retryAll` or `filter`
synchronously within the callback, and if they are not an
appropriate error will be thrown. This is mainly so that
the observer can be completed if the action is not going to
be retried, by checking after its next callback whether
`retryAll` or `filter` was called on the result.

Actions now also throw appropriate errors when `retryAll` or
`filter` is called more than once on the same result object,
or on a successful result object.

`Add` now calls `put` on retry, as retrying an add means that
the item already exists but we want to update it anyway.

Changed tests to reflect that:
 * Adding an existing item should be considered a data
   conflict, but the action should be able to be retried,
   and in that case it should update the existing item.
 * When subscribing to the store's observable, an initial
   update should be received that indicates the state of the
   store at the time of subscription.

Fixes two other miscellaneous bugs:
 * `_add` in `MemoryStore` was adding all passed items to
   the store's data. Now it correctly only adds only the new
   items.
 * On retry, the action was being executed immediately
   instead of being queued.